refactor(tests): Phase 1-2 of services test deduplication
Phase 1: Extract shared fixtures
- Extract stable_worktree_actor fixture to conftest.py (4 files)

Phase 2: Delete fragment files and trivial tests
- Delete test_issue_failure_refactored.py, merge edge case test
- Delete test_pr_create_usecase.py, merge to test_pr_service.py
- Delete 9 trivial tests across multiple files:
  * test_pr_service.py: 2 trivial tests
  * test_issue_title_cache_service.py: 3 trivial tests
  * test_issue_flow_service.py: 2 trivial tests
  * test_flow_orchestrator_service.py: 1 trivial test
  * test_capacity_service.py: 1 subsumed test
- Merge TestCleanResidualBranches from test_check_merge_fix.py to test_check_cleanup_service.py
