+++ tests/vibe3/services/test_check_cleanup_service.py
@@ -2,10 +2,38 @@
 
 from unittest.mock import MagicMock, patch
 
+import pytest
+
+from vibe3.clients import SQLiteClient
+from vibe3.clients.git_client import GitClient
 from vibe3.services.check_cleanup_service import CheckCleanupService
 from vibe3.services.expired_resource_cleanup_service import (
     ExpiredResourceCleanupService,
 )
+
+
+@pytest.fixture
+def mock_store():
+    """Create a mock SQLiteClient."""
+    return MagicMock(spec=SQLiteClient)
+
+
+@pytest.fixture
+def mock_git_client():
+    """Create a mock GitClient."""
+    client = MagicMock(spec=GitClient)
+    client.get_current_branch.return_value = "feature/test-branch"
+    client._run.return_value = "/path/to/.git"
+    return client
+
+
+@pytest.fixture
+def cleanup_service(mock_store, mock_git_client):
+    """Create a CheckCleanupService instance with mocked dependencies."""
+    return CheckCleanupService(
+        store=mock_store,
+        git_client=mock_git_client,
+    )
 
 
 def test_clean_residual_branches_filters_live_sessions_before_cleanup() -> None:
@@ -345,3 +373,110 @@ detached
         if c[0][0][:2] == ["worktree", "remove"]
     ]
     assert len(remove_calls) == 1
+
+
+class TestCleanResidualBranches:
+    """Tests for clean_residual_branches method in CheckCleanupService."""
+
+    def test_clean_residual_branches_removes_done_flow_branches(
+        self, cleanup_service, mock_store, mock_git_client
+    ):
+        """Should clean physical resources for done flows but keep flow record."""
+        mock_store.get_all_flows.return_value = [
+            {"branch": "feature/done-branch", "flow_status": "done"},
+            {"branch": "feature/active-branch", "flow_status": "active"},
+        ]
+        # Simulate local branch exists for done branch
+        mock_git_client._run.return_value = "feature/done-branch"
+        mock_git_client.find_worktree_path_for_branch.return_value = None
+
+        result = cleanup_service.clean_residual_branches()
+
+        # Done flow should be in kept_records (physical resources cleaned, record kept)
+        assert "feature/done-branch" in result["kept_records"]
+        assert len(result["cleaned"]) == 0  # cleaned is for aborted flows only
+        assert result["total_flows_checked"] == 1
+
+    def test_clean_residual_branches_skips_when_no_resources(
+        self, cleanup_service, mock_store, mock_git_client
+    ):
+        """Should preserve done/merged flow records even without physical resources.
+
+        Done/merged flows keep their records as completion history.
+        Only aborted flows have their records deleted.
+        """
+        mock_store.get_all_flows.return_value = [
+            {"branch": "feature/done-branch", "flow_status": "done"},
+        ]
+        # Simulate no resources exist
+        mock_git_client._run.return_value = ""
+        mock_git_client.find_worktree_path_for_branch.return_value = None
+
+        result = cleanup_service.clean_residual_branches()
+
+        # Done flow record should be preserved (in kept_records, not cleaned)
+        assert len(result["kept_records"]) == 1
+        assert "feature/done-branch" in result["kept_records"]
+        assert len(result["cleaned"]) == 0  # cleaned is for aborted flows
+        assert result["total_flows_checked"] == 1
+
+    def test_clean_residual_branches_aborted_deletes_record(
+        self, cleanup_service, mock_store, mock_git_client
+    ):
+        """Should delete aborted flow records to allow issue restart."""
+        mock_store.get_all_flows.return_value = [
+            {"branch": "feature/aborted-branch", "flow_status": "aborted"},
+        ]
+        # Simulate no resources exist
+        mock_git_client._run.return_value = ""
+        mock_git_client.find_worktree_path_for_branch.return_value = None
+
+        result = cleanup_service.clean_residual_branches()
+
+        # Aborted flow record should be deleted (in cleaned)
+        assert len(result["cleaned"]) == 1
+        assert "feature/aborted-branch" in result["cleaned"]
+        assert len(result["kept_records"]) == 0  # kept_records is for done/merged
+        assert result["total_flows_checked"] == 1
+
+    def test_clean_residual_branches_removes_invalid_records(
+        self, cleanup_service, mock_store, mock_git_client
+    ):
+        """Should remove HEAD records from database."""
+        mock_store.get_all_flows.return_value = [
+            {"branch": "HEAD", "flow_status": "done"},
+            {"branch": "HEAD~1", "flow_status": "aborted"},
+        ]
+
+        result = cleanup_service.clean_residual_branches()
+
+        assert len(result["removed_invalid"]) == 2
+        assert mock_store.delete_flow.call_count == 2
+
+    def test_clean_residual_branches_handles_partial_failures(
+        self, cleanup_service, mock_store, mock_git_client
+    ):
+        """Should continue cleaning even if some fail."""
+        mock_store.get_all_flows.return_value = [
+            {"branch": "feature/branch-1", "flow_status": "done"},
+            {"branch": "feature/branch-2", "flow_status": "done"},
+        ]
+        # First branch succeeds, second fails
+        call_count = [0]
+
+        def mock_run(cmd):
+            call_count[0] += 1
+            if call_count[0] <= 1:
+                return "feature/branch-1"  # First call for branch-1
+            elif call_count[0] <= 2:
+                raise RuntimeError("git error")  # First call for branch-2 fails
+            return ""  # Subsequent calls
+
+        mock_git_client._run.side_effect = mock_run
+        mock_git_client.find_worktree_path_for_branch.return_value = None
+
+        result = cleanup_service.clean_residual_branches()
+
+        # Should have some result despite failures
+        assert "failed" in result
+        assert result["total_flows_checked"] == 2
