+++ tests/vibe3/services/test_flow_main_branch_guard.py
@@ -8,19 +8,6 @@ from vibe3.models.flow import MainBranchProtectedError
 from vibe3.services.flow_service import FlowService
 
 
-@pytest.fixture(autouse=True)
-def stable_worktree_actor(monkeypatch):
-    """Avoid real git identity lookups during flow creation tests."""
-    monkeypatch.setattr(
-        "vibe3.services.flow_write_mixin.SignatureService.get_worktree_actor",
-        lambda: "test-actor",
-    )
-    monkeypatch.setattr(
-        "vibe3.services.flow_read_mixin.GitHubClient.get_pr",
-        lambda self, pr_number=None, branch=None: None,
-    )
-
-
 class TestMainBranchGuard:
     """Test main branch protection logic."""
